+++ scripts/build_ascent/build_ascent.sh
@@ -746,7 +746,7 @@ fi # build_vtkm
 ################
 # Camp
 ################
-camp_version=v2025.03.0
+camp_version=v2025.09.2
 camp_src_dir=$(ospath ${source_dir}/camp-${camp_version})
 camp_build_dir=$(ospath ${build_dir}/camp-${camp_version})
 camp_install_dir=$(ospath ${install_dir}/camp-${camp_version}/)
@@ -801,7 +801,7 @@ fi # build_camp
 ################
 # RAJA
 ################
-raja_version=v2025.03.2
+raja_version=v2025.09.0
 raja_src_dir=$(ospath ${source_dir}/RAJA-${raja_version})
 raja_build_dir=$(ospath ${build_dir}/raja-${raja_version})
 raja_install_dir=$(ospath ${install_dir}/raja-${raja_version}/)
@@ -860,7 +860,8 @@ fi # build_raja
 ################
 # Umpire
 ################
-umpire_version=2025.03.0
+# note: the release tarball naming scheme for Umpire is different vs RAJA + Camp
+umpire_version=2025.09.0
 umpire_src_dir=$(ospath ${source_dir}/umpire-${umpire_version})
 umpire_build_dir=$(ospath ${build_dir}/umpire-${umpire_version})
 umpire_install_dir=$(ospath ${install_dir}/umpire-${umpire_version}/)
@@ -1110,7 +1111,6 @@ fi
 
 if [[ "$enable_hip" == "ON" ]]; then
     echo 'set(ENABLE_HIP ON CACHE BOOL "")' >> ${root_dir}/ascent-config.cmake
-    echo 'set(BLT_CXX_STD c++17 CACHE STRING "")' >> ${root_dir}/ascent-config.cmake
     echo 'set(CMAKE_HIP_COMPILER ' ${CXX} ' CACHE STRING "")' >> ${root_dir}/ascent-config.cmake
     echo 'set(CMAKE_HIP_ARCHITECTURES ' ${ROCM_ARCH} ' CACHE STRING "")' >> ${root_dir}/ascent-config.cmake
     echo 'set(ROCM_PATH ' ${ROCM_PATH} ' CACHE PATH "")' >> ${root_dir}/ascent-config.cmake